fix lots compile warnings

+++ cocos/ui/UITextField.h
@@ -50,7 +50,7 @@ public:
     virtual bool onTextFieldAttachWithIME(TextFieldTTF *pSender) override;
     virtual bool onTextFieldDetachWithIME(TextFieldTTF * pSender) override;
     virtual bool onTextFieldInsertText(TextFieldTTF * pSender, const char * text, size_t nLen) override;
-    virtual bool onTextFieldDeleteBackward(TextFieldTTF * pSender, const char * delText, int nLen) override;
+    virtual bool onTextFieldDeleteBackward(TextFieldTTF * pSender, const char * delText, size_t nLen) override;
     
     void insertText(const char* text, size_t len);
     void deleteBackward();
